+++ api/core/src/modules/organization/organization.module.ts
@@ -1,10 +1,11 @@
 import { Module } from "@nestjs/common";
 import { OrganizationController } from "./organization.controller";
+import { OrganizationRepository } from "./organization.repository";
 import { OrganizationService } from "./organization.service";
 
 @Module({
   controllers: [OrganizationController],
-  providers: [OrganizationService],
+  providers: [OrganizationRepository, OrganizationService],
   exports: [OrganizationService]
 })
 export class OrganizationModule {}